Stubs for drawing hexes.
